\name{default_cd8_signature}
\alias{default_cd8_signature}
\title{Default CD8 Marker Signature}
\usage{
default_cd8_signature() 

}
\description{
Returns the default CD8 marker gene vector used by the CPTAC proteogenomic
validation wrapper when \code{signature_name = "CD8"}.
}
\value{
A character vector of gene symbols.
}
\examples{
default_cd8_signature()
}
