\name{default_treg_signature}
\alias{default_treg_signature}
\title{Default Treg signature genes}
\usage{
default_treg_signature() 

}
\description{
Returns the default Treg marker gene vector used by the pan-TCGA target-gene/signature survival workflow.
}
\value{
A character vector of gene symbols.
}
