% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_gene_plots.R
\name{plot_cell_gene_expression}
\alias{plot_cell_gene_expression}
\title{Plot tumor/reference donor expression}
\usage{
plot_cell_gene_expression(comparison)
}
\arguments{
\item{comparison}{Output of compare_cell_gene_expression.}
}
\value{
ggplot object; only donors used in the stated test are plotted.
}
\description{
Plot tumor/reference donor expression
}
